executor: migrate test-infra to testify for part of testSuite
