+++ executor/executor_legacy_test.go
@@ -56,7 +56,6 @@ import (
 	"github.com/pingcap/tidb/sessionctx"
 	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/store/copr"
-	error2 "github.com/pingcap/tidb/store/driver/error"
 	"github.com/pingcap/tidb/store/mockstore"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util"
@@ -1332,558 +1331,6 @@ func (s *testSuite2) TestStaleReadFutureTime(c *C) {
 	c.Assert(err, ErrorMatches, "cannot set read timestamp to a future time")
 	// TxnReadTS Is not updated if check failed.
 	c.Assert(tk.Se.GetSessionVars().TxnReadTS.PeakTxnReadTS(), Equals, uint64(0))
-}
-
-func (s *testSuite) TestScanControlSelection(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int primary key, b int, c int, index idx_b(b))")
-	tk.MustExec("insert into t values (1, 1, 1), (2, 1, 1), (3, 1, 2), (4, 2, 3)")
-	tk.MustQuery("select (select count(1) k from t s where s.b = t1.c) from t t1").Sort().Check(testkit.Rows("0", "1", "3", "3"))
-}
-
-func (s *testSuite) TestSimpleDAG(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t(a int primary key, b int, c int)")
-	tk.MustExec("insert into t values (1, 1, 1), (2, 1, 1), (3, 1, 2), (4, 2, 3)")
-	tk.MustQuery("select a from t").Check(testkit.Rows("1", "2", "3", "4"))
-	tk.MustQuery("select * from t where a = 4").Check(testkit.Rows("4 2 3"))
-	tk.MustQuery("select a from t limit 1").Check(testkit.Rows("1"))
-	tk.MustQuery("select a from t order by a desc").Check(testkit.Rows("4", "3", "2", "1"))
-	tk.MustQuery("select a from t order by a desc limit 1").Check(testkit.Rows("4"))
-	tk.MustQuery("select a from t order by b desc limit 1").Check(testkit.Rows("4"))
-	tk.MustQuery("select a from t where a < 3").Check(testkit.Rows("1", "2"))
-	tk.MustQuery("select a from t where b > 1").Check(testkit.Rows("4"))
-	tk.MustQuery("select a from t where b > 1 and a < 3").Check(testkit.Rows())
-	tk.MustQuery("select count(*) from t where b > 1 and a < 3").Check(testkit.Rows("0"))
-	tk.MustQuery("select count(*) from t").Check(testkit.Rows("4"))
-	tk.MustQuery("select count(*), c from t group by c order by c").Check(testkit.Rows("2 1", "1 2", "1 3"))
-	tk.MustQuery("select sum(c) as s from t group by b order by s").Check(testkit.Rows("3", "4"))
-	tk.MustQuery("select avg(a) as s from t group by b order by s").Check(testkit.Rows("2.0000", "4.0000"))
-	tk.MustQuery("select sum(distinct c) from t group by b").Check(testkit.Rows("3", "3"))
-
-	tk.MustExec("create index i on t(c,b)")
-	tk.MustQuery("select a from t where c = 1").Check(testkit.Rows("1", "2"))
-	tk.MustQuery("select a from t where c = 1 and a < 2").Check(testkit.Rows("1"))
-	tk.MustQuery("select a from t where c = 1 order by a limit 1").Check(testkit.Rows("1"))
-	tk.MustQuery("select count(*) from t where c = 1 ").Check(testkit.Rows("2"))
-	tk.MustExec("create index i1 on t(b)")
-	tk.MustQuery("select c from t where b = 2").Check(testkit.Rows("3"))
-	tk.MustQuery("select * from t where b = 2").Check(testkit.Rows("4 2 3"))
-	tk.MustQuery("select count(*) from t where b = 1").Check(testkit.Rows("3"))
-	tk.MustQuery("select * from t where b = 1 and a > 1 limit 1").Check(testkit.Rows("2 1 1"))
-
-	// Test time push down.
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (id int, c1 datetime);")
-	tk.MustExec("insert into t values (1, '2015-06-07 12:12:12')")
-	tk.MustQuery("select id from t where c1 = '2015-06-07 12:12:12'").Check(testkit.Rows("1"))
-
-	// Test issue 17816
-	tk.MustExec("drop table if exists t0")
-	tk.MustExec("CREATE TABLE t0(c0 INT)")
-	tk.MustExec("INSERT INTO t0 VALUES (100000)")
-	tk.MustQuery("SELECT * FROM t0 WHERE NOT SPACE(t0.c0)").Check(testkit.Rows("100000"))
-}
-
-func (s *testSuite) TestTimestampTimeZone(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (ts timestamp)")
-	tk.MustExec("set time_zone = '+00:00'")
-	tk.MustExec("insert into t values ('2017-04-27 22:40:42')")
-	// The timestamp will get different value if time_zone session variable changes.
-	tests := []struct {
-		timezone string
-		expect   string
-	}{
-		{"+10:00", "2017-04-28 08:40:42"},
-		{"-6:00", "2017-04-27 16:40:42"},
-	}
-	for _, tt := range tests {
-		tk.MustExec(fmt.Sprintf("set time_zone = '%s'", tt.timezone))
-		tk.MustQuery("select * from t").Check(testkit.Rows(tt.expect))
-	}
-
-	// For issue https://github.com/pingcap/tidb/issues/3467
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec(`CREATE TABLE t1 (
- 	      id bigint(20) NOT NULL AUTO_INCREMENT,
- 	      uid int(11) DEFAULT NULL,
- 	      datetime timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
- 	      ip varchar(128) DEFAULT NULL,
- 	    PRIMARY KEY (id),
- 	      KEY i_datetime (datetime),
- 	      KEY i_userid (uid)
- 	    );`)
-	tk.MustExec(`INSERT INTO t1 VALUES (123381351,1734,"2014-03-31 08:57:10","127.0.0.1");`)
-	r := tk.MustQuery("select datetime from t1;") // Cover TableReaderExec
-	r.Check(testkit.Rows("2014-03-31 08:57:10"))
-	r = tk.MustQuery("select datetime from t1 where datetime='2014-03-31 08:57:10';")
-	r.Check(testkit.Rows("2014-03-31 08:57:10")) // Cover IndexReaderExec
-	r = tk.MustQuery("select * from t1 where datetime='2014-03-31 08:57:10';")
-	r.Check(testkit.Rows("123381351 1734 2014-03-31 08:57:10 127.0.0.1")) // Cover IndexLookupExec
-
-	// For issue https://github.com/pingcap/tidb/issues/3485
-	tk.MustExec("set time_zone = 'Asia/Shanghai'")
-	tk.MustExec("drop table if exists t1")
-	tk.MustExec(`CREATE TABLE t1 (
-	    id bigint(20) NOT NULL AUTO_INCREMENT,
-	    datetime timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
-	    PRIMARY KEY (id)
-	  );`)
-	tk.MustExec(`INSERT INTO t1 VALUES (123381351,"2014-03-31 08:57:10");`)
-	r = tk.MustQuery(`select * from t1 where datetime="2014-03-31 08:57:10";`)
-	r.Check(testkit.Rows("123381351 2014-03-31 08:57:10"))
-	tk.MustExec(`alter table t1 add key i_datetime (datetime);`)
-	r = tk.MustQuery(`select * from t1 where datetime="2014-03-31 08:57:10";`)
-	r.Check(testkit.Rows("123381351 2014-03-31 08:57:10"))
-	r = tk.MustQuery(`select * from t1;`)
-	r.Check(testkit.Rows("123381351 2014-03-31 08:57:10"))
-	r = tk.MustQuery("select datetime from t1 where datetime='2014-03-31 08:57:10';")
-	r.Check(testkit.Rows("2014-03-31 08:57:10"))
-}
-
-func (s *testSuite) TestTimestampDefaultValueTimeZone(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("set time_zone = '+08:00'")
-	tk.MustExec(`create table t (a int, b timestamp default "2019-01-17 14:46:14")`)
-	tk.MustExec("insert into t set a=1")
-	r := tk.MustQuery(`show create table t`)
-	r.Check(testkit.Rows("t CREATE TABLE `t` (\n" + "  `a` int(11) DEFAULT NULL,\n" + "  `b` timestamp DEFAULT '2019-01-17 14:46:14'\n" + ") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin"))
-	tk.MustExec("set time_zone = '+00:00'")
-	tk.MustExec("insert into t set a=2")
-	r = tk.MustQuery(`show create table t`)
-	r.Check(testkit.Rows("t CREATE TABLE `t` (\n" + "  `a` int(11) DEFAULT NULL,\n" + "  `b` timestamp DEFAULT '2019-01-17 06:46:14'\n" + ") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin"))
-	r = tk.MustQuery(`select a,b from t order by a`)
-	r.Check(testkit.Rows("1 2019-01-17 06:46:14", "2 2019-01-17 06:46:14"))
-	// Test the column's version is greater than ColumnInfoVersion1.
-	sctx := tk.Se.(sessionctx.Context)
-	is := domain.GetDomain(sctx).InfoSchema()
-	c.Assert(is, NotNil)
-	tb, err := is.TableByName(model.NewCIStr("test"), model.NewCIStr("t"))
-	c.Assert(err, IsNil)
-	tb.Cols()[1].Version = model.ColumnInfoVersion1 + 1
-	tk.MustExec("insert into t set a=3")
-	r = tk.MustQuery(`select a,b from t order by a`)
-	r.Check(testkit.Rows("1 2019-01-17 06:46:14", "2 2019-01-17 06:46:14", "3 2019-01-17 06:46:14"))
-	tk.MustExec("delete from t where a=3")
-	// Change time zone back.
-	tk.MustExec("set time_zone = '+08:00'")
-	r = tk.MustQuery(`select a,b from t order by a`)
-	r.Check(testkit.Rows("1 2019-01-17 14:46:14", "2 2019-01-17 14:46:14"))
-	tk.MustExec("set time_zone = '-08:00'")
-	r = tk.MustQuery(`show create table t`)
-	r.Check(testkit.Rows("t CREATE TABLE `t` (\n" + "  `a` int(11) DEFAULT NULL,\n" + "  `b` timestamp DEFAULT '2019-01-16 22:46:14'\n" + ") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin"))
-
-	// test zero default value in multiple time zone.
-	defer tk.MustExec(fmt.Sprintf("set @@sql_mode='%s'", tk.MustQuery("select @@sql_mode").Rows()[0][0]))
-	tk.MustExec("set @@sql_mode='STRICT_TRANS_TABLES,NO_ENGINE_SUBSTITUTION';")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("set time_zone = '+08:00'")
-	tk.MustExec(`create table t (a int, b timestamp default "0000-00-00 00")`)
-	tk.MustExec("insert into t set a=1")
-	r = tk.MustQuery(`show create table t`)
-	r.Check(testkit.Rows("t CREATE TABLE `t` (\n" + "  `a` int(11) DEFAULT NULL,\n" + "  `b` timestamp DEFAULT '0000-00-00 00:00:00'\n" + ") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin"))
-	tk.MustExec("set time_zone = '+00:00'")
-	tk.MustExec("insert into t set a=2")
-	r = tk.MustQuery(`show create table t`)
-	r.Check(testkit.Rows("t CREATE TABLE `t` (\n" + "  `a` int(11) DEFAULT NULL,\n" + "  `b` timestamp DEFAULT '0000-00-00 00:00:00'\n" + ") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin"))
-	tk.MustExec("set time_zone = '-08:00'")
-	tk.MustExec("insert into t set a=3")
-	r = tk.MustQuery(`show create table t`)
-	r.Check(testkit.Rows("t CREATE TABLE `t` (\n" + "  `a` int(11) DEFAULT NULL,\n" + "  `b` timestamp DEFAULT '0000-00-00 00:00:00'\n" + ") ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_bin"))
-	r = tk.MustQuery(`select a,b from t order by a`)
-	r.Check(testkit.Rows("1 0000-00-00 00:00:00", "2 0000-00-00 00:00:00", "3 0000-00-00 00:00:00"))
-
-	// test add timestamp column default current_timestamp.
-	tk.MustExec(`drop table if exists t`)
-	tk.MustExec(`set time_zone = 'Asia/Shanghai'`)
-	tk.MustExec(`create table t (a int)`)
-	tk.MustExec(`insert into t set a=1`)
-	tk.MustExec(`alter table t add column b timestamp not null default current_timestamp;`)
-	timeIn8 := tk.MustQuery("select b from t").Rows()[0][0]
-	tk.MustExec(`set time_zone = '+00:00'`)
-	timeIn0 := tk.MustQuery("select b from t").Rows()[0][0]
-	c.Assert(timeIn8 != timeIn0, IsTrue, Commentf("%v == %v", timeIn8, timeIn0))
-	datumTimeIn8, err := expression.GetTimeValue(tk.Se, timeIn8, mysql.TypeTimestamp, 0)
-	c.Assert(err, IsNil)
-	tIn8To0 := datumTimeIn8.GetMysqlTime()
-	timeZoneIn8, err := time.LoadLocation("Asia/Shanghai")
-	c.Assert(err, IsNil)
-	err = tIn8To0.ConvertTimeZone(timeZoneIn8, time.UTC)
-	c.Assert(err, IsNil)
-	c.Assert(timeIn0 == tIn8To0.String(), IsTrue, Commentf("%v != %v", timeIn0, tIn8To0.String()))
-
-	// test add index.
-	tk.MustExec(`alter table t add index(b);`)
-	tk.MustExec("admin check table t")
-	tk.MustExec(`set time_zone = '+05:00'`)
-	tk.MustExec("admin check table t")
-
-	// 1. add a timestamp general column
-	// 2. add the index
-	tk.MustExec(`drop table if exists t`)
-	// change timezone
-	tk.MustExec(`set time_zone = 'Asia/Shanghai'`)
-	tk.MustExec(`create table t(a timestamp default current_timestamp)`)
-	tk.MustExec(`insert into t set a=now()`)
-	tk.MustExec(`alter table t add column b timestamp as (a+1) virtual;`)
-	// change timezone
-	tk.MustExec(`set time_zone = '+05:00'`)
-	tk.MustExec(`insert into t set a=now()`)
-	tk.MustExec(`alter table t add index(b);`)
-	tk.MustExec("admin check table t")
-	tk.MustExec(`set time_zone = '-03:00'`)
-	tk.MustExec("admin check table t")
-}
-
-func (s *testSuite) TestTiDBCurrentTS(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustQuery("select @@tidb_current_ts").Check(testkit.Rows("0"))
-	tk.MustExec("begin")
-	rows := tk.MustQuery("select @@tidb_current_ts").Rows()
-	tsStr := rows[0][0].(string)
-	txn, err := tk.Se.Txn(true)
-	c.Assert(err, IsNil)
-	c.Assert(tsStr, Equals, fmt.Sprintf("%d", txn.StartTS()))
-	tk.MustExec("begin")
-	rows = tk.MustQuery("select @@tidb_current_ts").Rows()
-	newTsStr := rows[0][0].(string)
-	txn, err = tk.Se.Txn(true)
-	c.Assert(err, IsNil)
-	c.Assert(newTsStr, Equals, fmt.Sprintf("%d", txn.StartTS()))
-	c.Assert(newTsStr, Not(Equals), tsStr)
-	tk.MustExec("commit")
-	tk.MustQuery("select @@tidb_current_ts").Check(testkit.Rows("0"))
-
-	_, err = tk.Exec("set @@tidb_current_ts = '1'")
-	c.Assert(terror.ErrorEqual(err, variable.ErrIncorrectScope), IsTrue, Commentf("err %v", err))
-}
-
-func (s *testSuite) TestTiDBLastTxnInfo(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (a int primary key)")
-	tk.MustQuery("select @@tidb_last_txn_info").Check(testkit.Rows(""))
-
-	tk.MustExec("insert into t values (1)")
-	rows1 := tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.start_ts'), json_extract(@@tidb_last_txn_info, '$.commit_ts')").Rows()
-	c.Assert(rows1[0][0].(string), Greater, "0")
-	c.Assert(rows1[0][0].(string), Less, rows1[0][1].(string))
-
-	tk.MustExec("begin")
-	tk.MustQuery("select a from t where a = 1").Check(testkit.Rows("1"))
-	rows2 := tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.start_ts'), json_extract(@@tidb_last_txn_info, '$.commit_ts'), @@tidb_current_ts").Rows()
-	tk.MustExec("commit")
-	rows3 := tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.start_ts'), json_extract(@@tidb_last_txn_info, '$.commit_ts')").Rows()
-	c.Assert(rows2[0][0], Equals, rows1[0][0])
-	c.Assert(rows2[0][1], Equals, rows1[0][1])
-	c.Assert(rows3[0][0], Equals, rows1[0][0])
-	c.Assert(rows3[0][1], Equals, rows1[0][1])
-	c.Assert(rows2[0][1], Less, rows2[0][2])
-
-	tk.MustExec("begin")
-	tk.MustExec("update t set a = a + 1 where a = 1")
-	rows4 := tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.start_ts'), json_extract(@@tidb_last_txn_info, '$.commit_ts'), @@tidb_current_ts").Rows()
-	tk.MustExec("commit")
-	rows5 := tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.start_ts'), json_extract(@@tidb_last_txn_info, '$.commit_ts')").Rows()
-	c.Assert(rows4[0][0], Equals, rows1[0][0])
-	c.Assert(rows4[0][1], Equals, rows1[0][1])
-	c.Assert(rows4[0][2], Equals, rows5[0][0])
-	c.Assert(rows4[0][1], Less, rows4[0][2])
-	c.Assert(rows4[0][2], Less, rows5[0][1])
-
-	tk.MustExec("begin")
-	tk.MustExec("update t set a = a + 1 where a = 2")
-	tk.MustExec("rollback")
-	rows6 := tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.start_ts'), json_extract(@@tidb_last_txn_info, '$.commit_ts')").Rows()
-	c.Assert(rows6[0][0], Equals, rows5[0][0])
-	c.Assert(rows6[0][1], Equals, rows5[0][1])
-
-	tk.MustExec("begin optimistic")
-	tk.MustExec("insert into t values (2)")
-	_, err := tk.Exec("commit")
-	c.Assert(err, NotNil)
-	rows7 := tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.start_ts'), json_extract(@@tidb_last_txn_info, '$.commit_ts'), json_extract(@@tidb_last_txn_info, '$.error')").Rows()
-	c.Assert(rows7[0][0], Greater, rows5[0][0])
-	c.Assert(rows7[0][1], Equals, "0")
-	c.Assert(strings.Contains(err.Error(), rows7[0][1].(string)), IsTrue)
-
-	_, err = tk.Exec("set @@tidb_last_txn_info = '{}'")
-	c.Assert(terror.ErrorEqual(err, variable.ErrIncorrectScope), IsTrue, Commentf("err %v", err))
-}
-
-func (s *testSerialSuite) TestTiDBLastTxnInfoCommitMode(c *C) {
-	defer config.RestoreFunc()()
-	config.UpdateGlobal(func(conf *config.Config) {
-		conf.TiKVClient.AsyncCommit.SafeWindow = time.Second
-	})
-
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (a int primary key, v int)")
-	tk.MustExec("insert into t values (1, 1)")
-
-	tk.MustExec("set @@tidb_enable_async_commit = 1")
-	tk.MustExec("set @@tidb_enable_1pc = 0")
-	tk.MustExec("update t set v = v + 1 where a = 1")
-	rows := tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.txn_commit_mode'), json_extract(@@tidb_last_txn_info, '$.async_commit_fallback'), json_extract(@@tidb_last_txn_info, '$.one_pc_fallback')").Rows()
-	c.Log(rows)
-	c.Assert(rows[0][0], Equals, `"async_commit"`)
-	c.Assert(rows[0][1], Equals, "false")
-	c.Assert(rows[0][2], Equals, "false")
-
-	tk.MustExec("set @@tidb_enable_async_commit = 0")
-	tk.MustExec("set @@tidb_enable_1pc = 1")
-	tk.MustExec("update t set v = v + 1 where a = 1")
-	rows = tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.txn_commit_mode'), json_extract(@@tidb_last_txn_info, '$.async_commit_fallback'), json_extract(@@tidb_last_txn_info, '$.one_pc_fallback')").Rows()
-	c.Assert(rows[0][0], Equals, `"1pc"`)
-	c.Assert(rows[0][1], Equals, "false")
-	c.Assert(rows[0][2], Equals, "false")
-
-	tk.MustExec("set @@tidb_enable_async_commit = 0")
-	tk.MustExec("set @@tidb_enable_1pc = 0")
-	tk.MustExec("update t set v = v + 1 where a = 1")
-	rows = tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.txn_commit_mode'), json_extract(@@tidb_last_txn_info, '$.async_commit_fallback'), json_extract(@@tidb_last_txn_info, '$.one_pc_fallback')").Rows()
-	c.Assert(rows[0][0], Equals, `"2pc"`)
-	c.Assert(rows[0][1], Equals, "false")
-	c.Assert(rows[0][2], Equals, "false")
-
-	c.Assert(failpoint.Enable("tikvclient/invalidMaxCommitTS", "return"), IsNil)
-	defer func() {
-		c.Assert(failpoint.Disable("tikvclient/invalidMaxCommitTS"), IsNil)
-	}()
-
-	tk.MustExec("set @@tidb_enable_async_commit = 1")
-	tk.MustExec("set @@tidb_enable_1pc = 0")
-	tk.MustExec("update t set v = v + 1 where a = 1")
-	rows = tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.txn_commit_mode'), json_extract(@@tidb_last_txn_info, '$.async_commit_fallback'), json_extract(@@tidb_last_txn_info, '$.one_pc_fallback')").Rows()
-	c.Log(rows)
-	c.Assert(rows[0][0], Equals, `"2pc"`)
-	c.Assert(rows[0][1], Equals, "true")
-	c.Assert(rows[0][2], Equals, "false")
-
-	tk.MustExec("set @@tidb_enable_async_commit = 0")
-	tk.MustExec("set @@tidb_enable_1pc = 1")
-	tk.MustExec("update t set v = v + 1 where a = 1")
-	rows = tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.txn_commit_mode'), json_extract(@@tidb_last_txn_info, '$.async_commit_fallback'), json_extract(@@tidb_last_txn_info, '$.one_pc_fallback')").Rows()
-	c.Log(rows)
-	c.Assert(rows[0][0], Equals, `"2pc"`)
-	c.Assert(rows[0][1], Equals, "false")
-	c.Assert(rows[0][2], Equals, "true")
-
-	tk.MustExec("set @@tidb_enable_async_commit = 1")
-	tk.MustExec("set @@tidb_enable_1pc = 1")
-	tk.MustExec("update t set v = v + 1 where a = 1")
-	rows = tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.txn_commit_mode'), json_extract(@@tidb_last_txn_info, '$.async_commit_fallback'), json_extract(@@tidb_last_txn_info, '$.one_pc_fallback')").Rows()
-	c.Log(rows)
-	c.Assert(rows[0][0], Equals, `"2pc"`)
-	c.Assert(rows[0][1], Equals, "true")
-	c.Assert(rows[0][2], Equals, "true")
-}
-
-func (s *testSuite) TestTiDBLastQueryInfo(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (a int primary key, v int)")
-	tk.MustQuery("select json_extract(@@tidb_last_query_info, '$.start_ts'), json_extract(@@tidb_last_query_info, '$.start_ts')").Check(testkit.Rows("0 0"))
-
-	toUint64 := func(str interface{}) uint64 {
-		res, err := strconv.ParseUint(str.(string), 10, 64)
-		c.Assert(err, IsNil)
-		return res
-	}
-
-	tk.MustExec("select * from t")
-	rows := tk.MustQuery("select json_extract(@@tidb_last_query_info, '$.start_ts'), json_extract(@@tidb_last_query_info, '$.for_update_ts')").Rows()
-	c.Assert(toUint64(rows[0][0]), Greater, uint64(0))
-	c.Assert(rows[0][0], Equals, rows[0][1])
-
-	tk.MustExec("insert into t values (1, 10)")
-	rows = tk.MustQuery("select json_extract(@@tidb_last_query_info, '$.start_ts'), json_extract(@@tidb_last_query_info, '$.for_update_ts')").Rows()
-	c.Assert(toUint64(rows[0][0]), Greater, uint64(0))
-	c.Assert(rows[0][0], Equals, rows[0][1])
-	// tidb_last_txn_info is still valid after checking query info.
-	rows = tk.MustQuery("select json_extract(@@tidb_last_txn_info, '$.start_ts'), json_extract(@@tidb_last_txn_info, '$.commit_ts')").Rows()
-	c.Assert(toUint64(rows[0][0]), Greater, uint64(0))
-	c.Assert(rows[0][0].(string), Less, rows[0][1].(string))
-
-	tk.MustExec("begin pessimistic")
-	tk.MustExec("select * from t")
-	rows = tk.MustQuery("select json_extract(@@tidb_last_query_info, '$.start_ts'), json_extract(@@tidb_last_query_info, '$.for_update_ts')").Rows()
-	c.Assert(toUint64(rows[0][0]), Greater, uint64(0))
-	c.Assert(rows[0][0], Equals, rows[0][1])
-
-	tk2 := testkit.NewTestKit(c, s.store)
-	tk2.MustExec("use test")
-	tk2.MustExec("update t set v = 11 where a = 1")
-
-	tk.MustExec("select * from t")
-	rows = tk.MustQuery("select json_extract(@@tidb_last_query_info, '$.start_ts'), json_extract(@@tidb_last_query_info, '$.for_update_ts')").Rows()
-	c.Assert(toUint64(rows[0][0]), Greater, uint64(0))
-	c.Assert(rows[0][0], Equals, rows[0][1])
-
-	tk.MustExec("update t set v = 12 where a = 1")
-	rows = tk.MustQuery("select json_extract(@@tidb_last_query_info, '$.start_ts'), json_extract(@@tidb_last_query_info, '$.for_update_ts')").Rows()
-	c.Assert(toUint64(rows[0][0]), Greater, uint64(0))
-	c.Assert(toUint64(rows[0][0]), Less, toUint64(rows[0][1]))
-
-	tk.MustExec("commit")
-
-	tk.MustExec("set transaction isolation level read committed")
-	tk.MustExec("begin pessimistic")
-	tk.MustExec("select * from t")
-	rows = tk.MustQuery("select json_extract(@@tidb_last_query_info, '$.start_ts'), json_extract(@@tidb_last_query_info, '$.for_update_ts')").Rows()
-	c.Assert(toUint64(rows[0][0]), Greater, uint64(0))
-	c.Assert(toUint64(rows[0][0]), Less, toUint64(rows[0][1]))
-
-	tk.MustExec("rollback")
-}
-
-func (s *testSuite) TestSelectForUpdate(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk1 := testkit.NewTestKit(c, s.store)
-	tk1.MustExec("use test")
-	tk2 := testkit.NewTestKit(c, s.store)
-	tk2.MustExec("use test")
-
-	tk.MustExec("drop table if exists t, t1")
-
-	txn, err := tk.Se.Txn(true)
-	c.Assert(kv.ErrInvalidTxn.Equal(err), IsTrue)
-	c.Assert(txn.Valid(), IsFalse)
-	tk.MustExec("create table t (c1 int, c2 int, c3 int)")
-	tk.MustExec("insert t values (11, 2, 3)")
-	tk.MustExec("insert t values (12, 2, 3)")
-	tk.MustExec("insert t values (13, 2, 3)")
-
-	tk.MustExec("create table t1 (c1 int)")
-	tk.MustExec("insert t1 values (11)")
-
-	// conflict
-	tk1.MustExec("begin")
-	tk1.MustQuery("select * from t where c1=11 for update")
-
-	tk2.MustExec("begin")
-	tk2.MustExec("update t set c2=211 where c1=11")
-	tk2.MustExec("commit")
-
-	_, err = tk1.Exec("commit")
-	c.Assert(err, NotNil)
-
-	// no conflict for subquery.
-	tk1.MustExec("begin")
-	tk1.MustQuery("select * from t where exists(select null from t1 where t1.c1=t.c1) for update")
-
-	tk2.MustExec("begin")
-	tk2.MustExec("update t set c2=211 where c1=12")
-	tk2.MustExec("commit")
-
-	tk1.MustExec("commit")
-
-	// not conflict
-	tk1.MustExec("begin")
-	tk1.MustQuery("select * from t where c1=11 for update")
-
-	tk2.MustExec("begin")
-	tk2.MustExec("update t set c2=22 where c1=12")
-	tk2.MustExec("commit")
-
-	tk1.MustExec("commit")
-
-	// not conflict, auto commit
-	tk1.MustExec("set @@autocommit=1;")
-	tk1.MustQuery("select * from t where c1=11 for update")
-
-	tk2.MustExec("begin")
-	tk2.MustExec("update t set c2=211 where c1=11")
-	tk2.MustExec("commit")
-
-	tk1.MustExec("commit")
-
-	// conflict
-	tk1.MustExec("begin")
-	tk1.MustQuery("select * from (select * from t for update) t join t1 for update")
-
-	tk2.MustExec("begin")
-	tk2.MustExec("update t1 set c1 = 13")
-	tk2.MustExec("commit")
-
-	_, err = tk1.Exec("commit")
-	c.Assert(err, NotNil)
-
-}
-
-func (s *testSuite) TestSelectForUpdateOf(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk1 := testkit.NewTestKit(c, s.store)
-	tk1.MustExec("use test")
-
-	tk.MustExec("drop table if exists t, t1")
-	tk.MustExec("create table t (i int)")
-	tk.MustExec("create table t1 (i int)")
-	tk.MustExec("insert t values (1)")
-	tk.MustExec("insert t1 values (1)")
-
-	tk.MustExec("begin pessimistic")
-	tk.MustQuery("select * from t, t1 where t.i = t1.i for update of t").Check(testkit.Rows("1 1"))
-
-	tk1.MustExec("begin pessimistic")
-
-	// no lock for t
-	tk1.MustQuery("select * from t1 for update").Check(testkit.Rows("1"))
-
-	// meet lock for t1
-	err := tk1.ExecToErr("select * from t for update nowait")
-	c.Assert(terror.ErrorEqual(err, error2.ErrLockAcquireFailAndNoWaitSet), IsTrue, Commentf("error: ", err))
-
-	// t1 rolled back, tk1 acquire the lock
-	tk.MustExec("rollback")
-	tk1.MustQuery("select * from t for update nowait").Check(testkit.Rows("1"))
-
-	tk1.MustExec("rollback")
-}
-
-func (s *testSuite) TestEmptyEnum(c *C) {
-	tk := testkit.NewTestKit(c, s.store)
-	tk.MustExec("use test")
-	tk.MustExec("drop table if exists t")
-	tk.MustExec("create table t (e enum('Y', 'N'))")
-	tk.MustExec("set sql_mode='STRICT_TRANS_TABLES'")
-	_, err := tk.Exec("insert into t values (0)")
-	c.Assert(terror.ErrorEqual(err, types.ErrTruncated), IsTrue, Commentf("err %v", err))
-	_, err = tk.Exec("insert into t values ('abc')")
-	c.Assert(terror.ErrorEqual(err, types.ErrTruncated), IsTrue, Commentf("err %v", err))
-
-	tk.MustExec("set sql_mode=''")
-	tk.MustExec("insert into t values (0)")
-	tk.MustQuery("select * from t").Check(testkit.Rows(""))
-	tk.MustExec("insert into t values ('abc')")
-	tk.MustQuery("select * from t").Check(testkit.Rows("", ""))
-	tk.MustExec("insert into t values (null)")
-	tk.MustQuery("select * from t").Check(testkit.Rows("", "", "<nil>"))
-
-	// Test https://github.com/pingcap/tidb/issues/29525.
-	tk.MustExec("drop table if exists t;")
-	tk.MustExec("create table t (id int auto_increment primary key, c1 enum('a', '', 'c'));")
-	tk.MustExec("insert into t(c1) values (0);")
-	tk.MustQuery("select id, c1+0, c1 from t;").Check(testkit.Rows("1 0 "))
-	tk.MustExec("alter table t change c1 c1 enum('a', '') not null;")
-	tk.MustQuery("select id, c1+0, c1 from t;").Check(testkit.Rows("1 0 "))
-	tk.MustExec("insert into t(c1) values (0);")
-	tk.MustQuery("select id, c1+0, c1 from t;").Check(testkit.Rows("1 0 ", "2 0 "))
 }
 
 const (
